+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## 2.74.3
+
+This release fixes a bug gracefully handling a request for `--validate-entry-point` when no
+`--entry-point` was given.
+
+* Error for missing entry point under `--validate-entry-point`. (#3048)
+
 ## 2.74.2
 
 This release fixes building PEXes from direct URL requirements. Previously, the direct URL